Add fetch accrual info

+++ internal/storage/postgres/postgres.go
@@ -23,7 +23,7 @@ func CreateOrdersTable(pool *pgxpool.Pool) error {
 		`CREATE TABLE IF NOT EXISTS orders (
 		id VARCHAR(255) PRIMARY KEY,
 		status VARCHAR(10) DEFAULT 'NEW' CHECK (status IN ('NEW', 'PROCESSING', 'INVALID', 'PROCESSED')),
-		accrual FLOAT DEFAULT 0.0,
+		accrual NUMERIC(10, 2) DEFAULT 0.0,
 		uploaded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 		user_id INTEGER NOT NULL REFERENCES users(id)
 	)`)
